### 2019-02-15 #1253 @whedon
[PRE REVIEW]: pyro: a framework for hydrodynamics explorations and prototyping
**Submitting author:** @harpolea (<a href="http://orcid.org/0000-0002-1530-781X">Alice Harpole</a>)
**Repository:** <a href="https://github.com/python-hydro/pyro2" target ="_blank">https://github.com/python-hydro/pyro2</a>
**Version:** 3.0
