### 2019-02-16 #1253 @labarba
It looks like something is failing with the embedded LaTeX symbol in this line:
 > method-of-lines piecewise linear solver$^\star$

Can you investigate, @harpolea? You could try `\ast` instead?

You can then recompile the paper yourself with the command `@whedon generate pdf` in a comment here.
