### 2019-02-17 #1253 @arfon
@harpolea - this PR should fix the paper compilation https://github.com/python-hydro/pyro2/pull/72
