### 2019-02-18 #1253 @mikaem
Looks interesting. I'll be happy to do it:-)
