### 2019-02-18 #1253 @labarba
@ngoldbaum Thank you! According to our policy, we simply ask for disclosure (as you've done here) of any association with the authors, and for an assessment on your part that the circumstance does not make you "unable to make an impartial scientific judgment or evaluation." If you assert that you can be impartial in your review, then we're good!
