### 2019-02-19 #1253 @ngoldbaum
I assert that my association with Mike doesn't make me unable to make an impartial scientific judgement or evaluation.
