### 2019-02-19 #1253 @gforsyth
> I wonder if you would be willing/able to contribute a review for JOSS on this submission? _"pyro: a framework for hydrodynamics explorations and prototyping"_

Happy to help -- currently slammed with SciPy stuff, but if you need another reviewer I can take a look next week
