### 2019-02-19 #1253 @labarba
Thanks for being willing to contribute a review for JOSS, @gforsyth! I will take you up on it another time, as we now have two reviewers for this submission and you're on SciPy duty.
