### 2019-02-20 python-hydro/pyro2#75 @ngoldbaum
Consider making pyro2 pip-installable
This is unrelated to my JOSS review so don't take this as a blocker for paper acceptance.

You may want to consider making pyro a pip-installable package, rather than relying on the directory structure of the git repository to make it runnable. This will make it easier for users to install pyro (people could just do `pip install pyro2` or `conda install -c conda-forge pyro2`) and also make it easier for developers who have worked with other python packages to get set up initially. If you did this you could make the `pyro.py` script an executable that gets installed to a user's UNIX `PATH` out of the box, so there would be no need to always work in the `pyro2` directory or refer to the `pyro.py` executable using its full path.
