### 2019-02-20 08c7b61
updated the README and installation instructions in the docs to mention that setting the PYTHONPATH is only necessary if using pyro as a python module
