### 2019-02-20 python-hydro/pyro2#75 @harpolea
I believe that the main reason pyro wasn't uploaded to pypi in the past was due to the f2py/Fortran dependence, however now that is no longer the case this is definitely something we'd like to do
