### 2019-02-20 9dc44a1
make the multigrid simple test script save a visualization by default
