### 2019-02-20 #1265 @ngoldbaum
I've gone through the reviewer checklist and I think this is ready to be published as-is.

My only substantive comment besides the issues and PRs I opened in the pyro2 repository is that the paper would be clearer if the third and second paragraph had their order switched (along with accompanying rewording so it flows well).
