### 2019-02-20 0a90bd2
switched paragraphs 2 and 3 in paper.md
