+++ paper/paper.md
@@ -37,20 +37,6 @@ implementation and exploration of hydrodynamics methods.  It is
 built in a object-oriented fashion, allowing for the reuse of
 the core components and fast prototyping of new methods.
 
-In the time since the first pyro paper [@pyroI], the code has
-undergone considerable development, gained a large number of solvers,
-adopted unit testing through pytest and documentation through sphinx,
-and a number of new contributors.  pyro's functionality can now
-be accessed directly through a `Pyro()` class, in addition to the
-original commandline script interface.  This new interface in particular
-allows for easy use within Jupyter notebooks.  We also now use HDF5
-for output instead of python's `pickle()` function.  Previously, we used Fortran
-to speed up some performance-critical portions of the code.  These routines
-could be called by the main python code by first compiling them using `f2py`.
-In the new version, we have replaced these Fortran routines by python functions
-that are compiled at runtime by `numba`.  Consequently, pyro is now written
-entirely in python.
-
 The original goal of pyro was to learn hydrodynamics methods through
 example, and it still serves this goal.  At Stony Brook, pyro is used
 with new undergraduate researchers in our group to introduce them to
@@ -65,6 +51,20 @@ development of higher-order schemes in the Castro hydrodynamics code
 on the Maestro code [@maestro] and the pyro implementation will be
 used to prototype new low Mach number algorithms before porting them
 to science codes.
+
+In the time since the first pyro paper [@pyroI], the code has
+undergone considerable development, gained a large number of solvers,
+adopted unit testing through pytest and documentation through sphinx,
+and a number of new contributors.  pyro's functionality can now
+be accessed directly through a `Pyro()` class, in addition to the
+original commandline script interface.  This new interface in particular
+allows for easy use within Jupyter notebooks.  We also now use HDF5
+for output instead of python's `pickle()` function.  Previously, we used Fortran
+to speed up some performance-critical portions of the code.  These routines
+could be called by the main python code by first compiling them using `f2py`.
+In the new version, we have replaced these Fortran routines by python functions
+that are compiled at runtime by `numba`.  Consequently, pyro is now written
+entirely in python.
 
 The current pyro solvers are:
 
