### 2019-02-20 #1265 @harpolea
Thanks @ngoldbaum! I agree about the paragraph ordering - I've switched them over in python-hydro/pyro2@0a90bd25fc9f750370847f3cd6e0a51f82f78f5d
