### 2019-02-21 #1265 @mikaem
Hi
Sorry to hold up the review process. I've been familiarising myself with the project, and I started by following the installation instructions on readthedocs. No problem thus far and I get a nice figure from running

`./pyro.py advection smooth inputs.smooth`

All in all a very nice project:-) I think all there is for me to do really is to tick off boxes and provide just a few advices on improvement (having read the JOSS Guiding priciples: 'a key goal of JOSS is to raise the quality of research software'). And this is not meant as lengthy details of difficulties, just some comments.

I've seen that @ngoldbaum has opened up an issue https://github.com/python-hydro/pyro2/issues/75 regarding installation and I strongly support this suggestion. Furthermore, a good structure with a base solver module and a base problem module would probably make the code easier to follow, and reuse of code would be more intuitive. The softlinked problem folders inside many of the solver folders is probably not the best design;-)

Regarding the documentation, I think a lot of double ``quotes`` could be replaced with appropriate cross-links that would make the documentation easier to follow. (For example when the [compare.py](https://github.com/python-hydro/pyro2/blob/master/compare.py) utility is [documented](https://pyro2.readthedocs.io/en/latest/output.html) and many others like it.) The API documentation is nicely done (for most) using Numpy style. However, the built API documentation looks a bit strange [(see e.g., Parameters here)](https://pyro2.readthedocs.io/en/latest/advection.html) and I think this is because you are missing 'sphinx.ext.napoleon' from the list of extensions in [conf.py](https://github.com/python-hydro/pyro2/blob/94ab173f40b8ce752bc3392b42c9e63e7cc818b3/docs/source/conf.py#L35). See [also here](http://www.sphinx-doc.org/en/master/usage/extensions/napoleon.html)
Also, from the documentation it is not really clear to me how to contribute to this project, only how to get help. May I suggest (as I should probably have done in an issue), that you add a small paragraph around [here](https://pyro2.readthedocs.io/en/latest/help.html) about ``how to contribute``.

Just a few minor comments to a very nice project. I'll go ahead now and see if I can tick off some boxes:-)
