### 2019-02-21 1f14c58
added contributing guide to docs
