+++ docs/source/help.rst
@@ -1,5 +1,44 @@
-Getting help
-============
+Contributing and getting help
+=============================
+
+Contributing
+------------
+
+Contributions are welcomed from anyone, including posting issues or
+submitting pull requests to the [pyro github](https://github.com/python-hydro/pyro).
+
+Users who make significant contributions will be listed as developers
+in the pyro acknowledgements and be included in any future code
+papers.
+
+Issues
+------
+
+Creating an issue on github is a good way to request new features,
+file a bug report, or notify us of any difficulties that arise using
+pyro.
+
+To request support using pyro, please create an issue on the pyro
+github and the developers will be happy to assist you.
+
+If you are reporting a bug, please indicate any information necessary
+to reproduce the bug including your version of python.
+
+Pull Requests
+-------------
+
+*Any contributions that have the potential to change answers should be
+done via pull requests.* A pull request should be generated from your
+fork of pyro and target the `master` branch.
+
+The unit and regression tests will run automatically once the PR is
+submitted, and then one of the pyro developers will review the PR and
+if needed, suggest modifications prior to merging the PR.
+
+If there are a number of small commits making up the PR, we may wish
+to squash commits upon merge to have a clean history.  *Please ensure
+that your PR title and first post are descriptive, since these will be
+used for a squashed commit message.*
 
 Mailing list
 ------------
@@ -8,6 +47,3 @@ There is a public mailing list for discussing pyro. Visit the
 `pyro-code@googlegroups.com
 <https://groups.google.com/forum/#!forum/pyro-code>`_ and
 subscribe. You can then send questions to that list.
-
-
-
