### 2019-02-21 #1265 @harpolea
Thanks for your comments @mikaem!

- Currently `simulation_null.py` acts as a base for the solvers. As the solvers and problems can be quite different from each other (e.g. compare compressible vs. advection), I'm not really sure that there would be much benefit in having e.g. additional (abstract) base classes for the solvers and problems, and that it may just add an additional level of complexity.
- I agree that more links in the documentation would be helpful - I'll add those in now
- The docstrings in the API are currently processed using the `numpydoc` extension. The only difference I see between that and `sphinx.ext.napoleon` seems to be formatting (i.e. the former highlights the variable names, whereas the latter does not)?
- We have a `CONTRIBUTING.md` file in the main repo, but I agree that this should be added to the documentation as well. I've added that in python-hydro/pyro2@1f14c583
