+++ docs/source/installation.rst
@@ -21,7 +21,7 @@ The following steps are needed before running pyro:
 * add ``pyro/`` to your ``PYTHONPATH`` environment variable (note this is only
 needed if you wish to use pyro as a python
 module - this step is not necessary if you only run pyro via the
-commandline using the `pyro.py` script).  For
+commandline using the ``pyro.py`` script).  For
   the bash shell, this is done as:
 
     .. code-block:: none
