### 2019-02-21 #1265 @harpolea
I added some more links to the documentation in python-hydro/pyro2@47174438. In the example you linked to, `compare.py` refers to the script itself rather than the module it contains. I think that linking to the module here would not be particularly helpful, as the function of that part of the documentation is to document the use of the script itself.
