### 2019-02-21 6c531f0
fixed link in help.rst
