### 2019-02-21 #1265 @mikaem
> Thanks for your comments @mikaem!
>
> * Currently `simulation_null.py` acts as a base for the solvers. As the solvers and problems can be quite different from each other (e.g. compare compressible vs. advection), I'm not really sure that there would be much benefit in having e.g. additional (abstract) base classes for the solvers and problems, and that it may just add an additional level of complexity.

I understand completely. It's just that when I see a lot of code duplication (or even soft links) I immediately think that this could probably be solved with object orientation and overloading. For example, I could imagine a problem base module where a default `init_data` method was defined, like the one under [advection/problems/smooth.py](https://github.com/python-hydro/pyro2/blob/master/advection/problems/smooth.py). Then the problem under advection_rk/problems could simply import that function instead of the soft link. And the problem that was using an advection_fv4 solver could import the base method and overload it appropriately. Works similarly if there was a base problem class with an init_data method to overload. Anyways, how you do it is entirely up to you. I'm not going to fuss over it.

> * I agree that more links in the documentation would be helpful - I'll add those in now
> * The docstrings in the API are currently processed using the `numpydoc` extension. The only difference I see between that and `sphinx.ext.napoleon` seems to be formatting (i.e. the former highlights the variable names, whereas the latter does not)?

Below I add two links to figures that are compiling your documentation with and without ``sphinx.ext.napoleon``. I would say that the second looks much nicer, but it is only formatting, you are right about that:-)

[Without sphinx.ext.napoleon, i.e., from readthedocs](https://www.dropbox.com/s/72tlc305mw4esql/Screenshot_without_napoleon.png?dl=0)
[With sphinx.ext.napoleon, compiled locally](https://www.dropbox.com/s/xm1qdyh782ioo6e/Screenshot_with_napoleon.png?dl=0)

> * We have a `CONTRIBUTING.md` file in the main repo, but I agree that this should be added to the documentation as well. I've added that in [python-hydro/pyro2@1f14c58](https://github.com/python-hydro/pyro2/commit/1f14c583)

Ok, great. In that case I'll tick off my last box and I'm ready to sign off on this:-)
