### 2019-02-21 #1265 @labarba
Minor editorial suggestions:

- Perhaps write `pyro` in code format? (otherwise, it looks weird in lower case starting a sentence)
- pyro is a python-based >> capitalize Python
- The low Mach number atmospheric solver >> low-Mach-number (hyphenate compound adjective)
(twice)
- python's `pickle()` >> capitalize Python
- main python code >> capitalize Python
- python functions >> idem
- entirely in python >> idem
- low Mach number atmospheric solver* >> low-Mach-number
- Also new support >> comma after "Also"
