### 2019-02-21 #1265 @harpolea
I've updated the paper file in python-hydro/pyro2@1ab2230a to capitalize the Python's and format the `pyro`'s. In the (astrophysical) literature, it's normally written as low Mach number without the hyphens, so I think it would be better left as it is?
