### 2019-02-22 #1265 @labarba
Hi @harpolea — Time to make a tagged release of the software (and report the version number here), then an archive in Zenodo or a similar service (and report the DOI here). Cheers!
