### 2019-02-22 #1265 @harpolea
Hi @labarba - I've made a tagged release (v. 3.1) and have created a Zenodo DOI ([10.5281/zenodo.2575565](https://doi.org/10.5281/zenodo.2575565))
