### 2019-02-22 #1265 @labarba
Can you edit the metadata of the Zenodo deposit so it matches the title and author list of the JOSS paper?
