### 2019-02-22 #1265 @labarba
Congratulations, @harpolea et al., your paper is published in JOSS!

Many thanks to our reviewers, @mikaem, @ngoldbaum—you make this possible 🙏
